Changed LinqTranslatorVisitor.cs into a partial class and moved a lot of method translators into separate files of the same class.
