+++ src/LinqSql/Visitors/LinqTranslatorVisitor/LinqTranslatorVisitor.cs
@@ -1,13 +1,11 @@
-﻿using System.Collections.Generic;
-using System.Reflection;
-using System.Linq.Expressions;
+﻿using System.Linq.Expressions;
 
 namespace System.Linq.Sql
 {
     /// <summary>
     /// <see cref="LinqTranslatorVisitor"/> translates an expression call to an sql expression tree.
     /// </summary>
-    public class LinqTranslatorVisitor : ExpressionVisitor
+    public partial class LinqTranslatorVisitor : ExpressionVisitor
     {
         private ASourceExpression[] sources = null;
 
@@ -41,7 +39,7 @@ namespace System.Linq.Sql
                 return base.Visit(node);
         }
 
-        private T Visit<T>(Expression expression)
+        internal T Visit<T>(Expression expression)
             where T : AExpression
         {
             T source = Visit(StripQuotes(expression)) as T;
@@ -185,307 +183,6 @@ namespace System.Linq.Sql
                 default:
                     throw new MethodTranslationException(node.Method);
             }
-        }
-
-        private ContainsExpression VisitContains(MethodCallExpression expression)
-        {
-            // Handle extension methods defined by Linqs
-            if (expression.Method.DeclaringType == typeof(Enumerable) || expression.Method.DeclaringType == typeof(Queryable))
-            {
-                AExpression values = Visit<AExpression>(expression.Arguments[0]);
-                AExpression value = Visit<AExpression>(expression.Arguments[1]);
-                return new ContainsExpression(values, value);
-            }
-
-            // Handle custom extension method
-            if (expression.Method.DeclaringType == typeof(SqlQueryableHelper))
-            {
-                // Get value expression first, because the source will change to the subquery making the value out of scope
-                AExpression value = Visit<AExpression>(expression.Arguments[2]);
-
-                // Evaluate the subquery expressions
-                ASourceExpression source = Visit<ASourceExpression>(expression.Arguments[0]);
-                LambdaExpression fieldLambda = (LambdaExpression)StripQuotes(expression.Arguments[1]);
-                FieldExpression field = Visit<FieldExpression>(fieldLambda.Body);
-
-                // Create the expression
-                return new ContainsExpression(new ScalarExpression(source, field), value);
-            }
-
-            throw new MethodTranslationException(expression.Method);
-        }
-
-        private AggregateExpression VisitAggregate(MethodCallExpression expression, AggregateFunction function)
-        {
-            Type type = expression.Method.DeclaringType;
-            if (type == typeof(SqlQueryableHelper) || type == typeof(Enumerable) || type == typeof(Queryable))
-            {
-                // Resolve the source
-                ASourceExpression source = Visit<ASourceExpression>(expression.Arguments[0]);
-
-                // Resolve the optional selector
-                if (expression.Arguments.Count > 1)
-                {
-                    LambdaExpression lambda = (LambdaExpression)StripQuotes(expression.Arguments[1]);
-                    FieldExpression found = Visit<FieldExpression>(lambda.Body);
-                    source = new SelectExpression(source, new[] { found });
-                }
-
-                // Resolve the field to be counted (must be done after the source has been manipulated)
-                FieldExpression field = source.Fields.First();
-
-                // Create the expression
-                return new AggregateExpression(source, field, function);
-            }
-
-            throw new MethodTranslationException(expression.Method);
-        }
-
-        private AggregateExpression VisitCount(MethodCallExpression expression)
-        {
-            Type type = expression.Method.DeclaringType;
-            if (type == typeof(SqlQueryableHelper) || type == typeof(Enumerable) || type == typeof(Queryable))
-            {
-                // Map the source
-                ASourceExpression source = Visit<ASourceExpression>(expression.Arguments[0]);
-
-                // Resolve the optional predicate
-                if (expression.Arguments.Count > 1)
-                {
-                    LambdaExpression lambda = (LambdaExpression)StripQuotes(expression.Arguments[1]);
-                    APredicateExpression predicate = Visit<APredicateExpression>(lambda.Body);
-                    source = new WhereExpression(source, predicate);
-                }
-
-                // Resolve the field to be counted (must be done after the source has been manipulated)
-                FieldExpression field = source.Fields.First();
-
-                // Create the expression
-                return new AggregateExpression(source, field, AggregateFunction.Count);
-
-            }
-
-            throw new MethodTranslationException(expression.Method);
-        }
-
-        private SelectExpression VisitSkip(MethodCallExpression expression)
-        {
-            if (expression.Method.DeclaringType == typeof(Enumerable) || expression.Method.DeclaringType == typeof(Queryable))
-            {
-                ASourceExpression source = Visit<ASourceExpression>(expression.Arguments[0]);
-                int count = (int)((ConstantExpression)expression.Arguments[1]).Value;
-                return new SelectExpression(source, source.Fields, -1, count);
-            }
-
-            throw new MethodTranslationException(expression.Method);
-        }
-
-        private SelectExpression VisitTake(MethodCallExpression expression)
-        {
-            if (expression.Method.DeclaringType == typeof(Enumerable) || expression.Method.DeclaringType == typeof(Queryable))
-            {
-                ASourceExpression source = Visit<ASourceExpression>(expression.Arguments[0]);
-                int count = (int)((ConstantExpression)expression.Arguments[1]).Value;
-                return new SelectExpression(source, source.Fields, count, 0);
-            }
-
-            throw new MethodTranslationException(expression.Method);
-        }
-
-        private SelectExpression VisitOrderBy(MethodCallExpression expression)
-        {
-            MethodInfo method = expression.Method;
-            Type type = method.DeclaringType;
-            if (type == typeof(SqlQueryableHelper) || type == typeof(Enumerable) || type == typeof(Queryable))
-            {
-                // Resolve the source
-                ASourceExpression source = Visit<ASourceExpression>(expression.Arguments[0]);
-
-                // Resolve the optional selector
-                FieldExpression field = source.Fields.First();
-                if (expression.Arguments.Count > 1)
-                {
-                    LambdaExpression lambda = (LambdaExpression)StripQuotes(expression.Arguments[1]);
-                    field = Visit<FieldExpression>(lambda.Body);
-                }
-
-                // Decode the direction
-                OrderType direction = method.Name.EndsWith("Descending") ? OrderType.Descending : OrderType.Ascending;
-
-                // Handle an existing select expression
-                if (source is SelectExpression select)
-                {
-                    if (method.Name.StartsWith("ThenBy") && !select.Orderings.Any())
-                        throw new InvalidOperationException($"{method.Name} can only be applied to an ordered sequence.");
-                    if (method.Name.StartsWith("OrderBy") && select.Orderings.Any())
-                        throw new InvalidOperationException($"{method.Name} can only be applied to an unordered sequence.");
-
-                    // Clone and modify the select expression
-                    IEnumerable<FieldExpression> fields = select.Fields.Select(x => x.SourceExpression);
-                    IEnumerable<Ordering> orderings = select.Orderings.Concat(new[] { new Ordering(field.SourceExpression, direction) });
-                    return new SelectExpression(select.Source, fields, select.Take, select.Skip, orderings);
-                }
-
-                // Create the expression
-                if (method.Name.StartsWith("ThenBy"))
-                    throw new InvalidOperationException($"{method.Name} can only be applied to an ordered sequence.");
-                return new SelectExpression(source, orderings: new[] { new Ordering(field, direction) });
-            }
-
-            throw new MethodTranslationException(expression.Method);
-        }
-
-        private WhereExpression VisitWhere(MethodCallExpression expression)
-        {
-            // Handle the default Queryable extension Where
-            if (expression.Method.DeclaringType == typeof(Queryable))
-            {
-                ASourceExpression source = Visit<ASourceExpression>(expression.Arguments[0]);
-                LambdaExpression lambda = (LambdaExpression)StripQuotes(expression.Arguments[1]);
-                APredicateExpression predicate = Visit<APredicateExpression>(lambda.Body);
-                return new WhereExpression(source, predicate);
-            }
-
-            throw new MethodTranslationException(expression.Method);
-        }
-
-        private JoinExpression VisitJoin(MethodCallExpression expression)
-        {
-            // Handle the default Queryable extension Join
-            if (expression.Method.DeclaringType == typeof(Queryable))
-            {
-                // Resolve the sources
-                ASourceExpression outer = Visit<ASourceExpression>(expression.Arguments[0]);
-                ASourceExpression inner = Visit<ASourceExpression>(expression.Arguments[1]);
-
-                // Set the active expressions (so fields calls can find their expression)
-                sources = new[] { outer, inner };
-
-                // Create the predicate
-                LambdaExpression outerLambda = (LambdaExpression)StripQuotes(expression.Arguments[2]);
-                LambdaExpression innerLambda = (LambdaExpression)StripQuotes(expression.Arguments[3]);
-                FieldExpression outerField = Visit<FieldExpression>(outerLambda.Body);
-                FieldExpression innerField = Visit<FieldExpression>(innerLambda.Body);
-                APredicateExpression predicate = new CompositeExpression(outerField, innerField, CompositeOperator.Equal);
-
-                // Decode the result selector
-                IEnumerable<FieldExpression> fields = DecodeJoinSelector(expression.Arguments[4], outer.Fields, inner.Fields);
-
-                // Create the expression
-                return new JoinExpression(outer, inner, predicate, fields, JoinType.Inner);
-            }
-
-            // Handle the default SqlQueryableHelper extension Join
-            if (expression.Method.DeclaringType == typeof(SqlQueryableHelper))
-            {
-                // Resolve the sources
-                ASourceExpression outer = Visit<ASourceExpression>(expression.Arguments[0]);
-                ASourceExpression inner = Visit<ASourceExpression>(expression.Arguments[1]);
-
-                // Set the active expressions (so fields calls can find their expression)
-                sources = new[] { outer, inner };
-
-                // Create the predicate
-                LambdaExpression predicateLambda = (LambdaExpression)StripQuotes(expression.Arguments[2]);
-                APredicateExpression predicate = Visit<APredicateExpression>(predicateLambda.Body);
-
-                // Decode the result selector
-                IEnumerable<FieldExpression> fields = DecodeJoinSelector(expression.Arguments[3], outer.Fields, inner.Fields);
-
-                // Resolve the join type
-                ConstantExpression joinType = (ConstantExpression)expression.Arguments[4];
-
-                // Create the expression
-                return new JoinExpression(outer, inner, predicate, fields, (JoinType)joinType.Value);
-            }
-
-            throw new MethodTranslationException(expression.Method);
-        }
-
-        private IEnumerable<FieldExpression> DecodeJoinSelector(Expression expression, FieldExpressions outer, FieldExpressions inner)
-        {
-            // Get the associated lambda
-            LambdaExpression lambda = StripQuotes(expression) as LambdaExpression;
-
-            // Decode paramter body
-            ParameterExpression parameter = lambda.Body as ParameterExpression;
-            if (parameter != null)
-            {
-                if (parameter == lambda.Parameters[0])
-                    return outer;
-                else if (parameter == lambda.Parameters[1])
-                    return inner;
-
-                throw new InvalidOperationException($"The Join result selector does not recognise the specified '{parameter}' parameter. When using a paramter it must be one of the result selector paramter list.");
-            }
-
-            // Decode binary selector (It should always evaluate to all fields)
-            BinaryExpression binary = lambda.Body as BinaryExpression;
-            if (binary != null)
-            {
-                // Check the correct operator is being used
-                if (binary.NodeType != ExpressionType.Or)
-                    throw new NotSupportedException("The only supported binary operator of an Sql Join call is Or: '|'.");
-
-                // Check the binary expressions are the paramters specified in the expression
-                if (binary.Left != lambda.Parameters[0] && binary.Left != lambda.Parameters[1])
-                    throw new NotSupportedException("The left of the '|' operator in a Join result selector must be a parameter of the selector expression.");
-                if (binary.Right != lambda.Parameters[0] && binary.Right != lambda.Parameters[1])
-                    throw new NotSupportedException("The right of the '|' operator in a Join result selector must be a parameter of the selector expression.");
-
-                // Resolve select all
-                if ((binary.Left == lambda.Parameters[0] && binary.Right == lambda.Parameters[1]) ||
-                    (binary.Left == lambda.Parameters[1] && binary.Right == lambda.Parameters[0]))
-                    return null;
-
-                throw new InvalidOperationException("The Join result selector should not select the same parameter more than once. e.g. '(o, i) => i | i' should be replaced with '(o, i) => i'.");
-            }
-
-            throw new NotSupportedException($"The Join result selctor does not know how to translate '{expression}' to SQL. The supported expressions are: '(o, i) => o', '(o, i) => i', '(o, i) => o | i' and '(o, i) => i | o'.");
-        }
-
-        private FieldExpression VisitField(MethodCallExpression expression)
-        {
-            if (sources == null)
-                throw new InvalidOperationException($"A field can only be visited if an {nameof(ASourceExpression)} has previously been visited.");
-
-            // Resolve the field name
-            ConstantExpression fieldNameExpression = expression.Arguments.FirstOrDefault() as ConstantExpression;
-            if (!expression.Method.DeclaringType.IsAssignableFrom(typeof(Dictionary<string, object>)))
-                throw new InvalidOperationException("The declaring type for a field expression must be a Dictionary<string, object>.");
-            if (expression.Method.ReturnType != typeof(object))
-                throw new InvalidOperationException("The return type for a field expression must be type of object.");
-            if (expression.Arguments.Count != 1 || fieldNameExpression?.Type != typeof(string))
-                throw new InvalidOperationException("The field name indexer for the field expression must contain exactly one string parameter.");
-
-            // Resolve the table name
-            MethodCallExpression source = expression.Object as MethodCallExpression;
-            ConstantExpression tableNameExpression = source?.Arguments.FirstOrDefault() as ConstantExpression;
-            if (source == null)
-                throw new InvalidOperationException($"The table instance object could not be resolved for the field: {fieldNameExpression.Value.ToString()}");
-            if (source.Method.Name != "get_Item")
-                throw new NotSupportedException("Only an array indexer can be used to resolve a fields table name.");
-            if (!source.Method.ReturnType.IsAssignableFrom(typeof(RecordItem)))
-                throw new InvalidOperationException($"When mapping a field, the table name must map to a {nameof(RecordItem)}.");
-            if (source.Arguments.Count != 1 || tableNameExpression?.Type != typeof(string))
-                throw new InvalidOperationException("The table name indexer for the field expression must contain exactly one string parameter.");
-
-            // Resolve the indexer values
-            string tableName = tableNameExpression.Value as string;
-            string fieldName = fieldNameExpression.Value as string;
-            if (string.IsNullOrWhiteSpace(tableName))
-                throw new InvalidOperationException("The table name cannot be empty.");
-            if (string.IsNullOrWhiteSpace(fieldName))
-                throw new InvalidOperationException("The field name cannot be empty.");
-
-            // Get the field from the current source
-            FieldExpression found = sources?
-                .SelectMany(x => x.Fields)
-                .FirstOrDefault(x => x.TableName == tableName && x.FieldName == fieldName);
-            if (found == null)
-                throw new KeyNotFoundException($"The field [{tableName}].[{fieldName}] could not be found on the current source expression.");
-
-            return found;
         }
 
         private static Expression StripQuotes(Expression expression)
